pyproject.toml
[project]
name = "rag"
version = "0.2.0"
description = "Agentic RAG Pipeline — LangGraph ReAct agent with tool-calling"
requires-python = ">=3.11"
dependencies = [
    # LangChain core
    "langchain",
    "langchain-community",
    "langchain-core",
    "langchain-text-splitters",
    "langchain-experimental",
    "langchain-qdrant",
    "langchain-huggingface",
    # Agent framework
    "langgraph",
    # Web search tool
    "tavily-python",
    # Vector Store
    "qdrant-client>=1.10.0",
    # PDF Processing
    "pymupdf>=1.24.0",
    "pypdf>=4.0.0",
    # Embeddings (HuggingFace)
    "sentence-transformers>=2.6.0",
    "transformers>=4.38.0",
    "huggingface-hub>=0.34.0,<1.0",
    "torch>=2.2.0",
    "accelerate>=0.26.0",
    "sentencepiece>=0.2.0",
    "safetensors>=0.4.2",
    # LLM Backend
    "vllm>=0.6.0",
    "langchain-openai>=0.3.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "cachetools>=5.3.0",
    "rank-bm25>=0.2.2",
    "numpy>=1.24.0,<2.3",
    "scikit-learn>=1.3.0",
    # API Gateway
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
    # Database (Summary Store)
    "psycopg2-binary>=2.9.0",
    # Multi-Provider LLM Router (optional)
    "litellm>=1.40.0",
    # Redis
    "redis>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
